Add support for 'user_object_parsers' option to ExMarshal.decode/2
This option can be used to specify a map associating atoms with unary
functions. These functions will be considered when trying to decode user
objects. I.e. instead of user objects resulting in 2-tuples {class_name,
attributes}, a custom function is invoked which can then do something
interesting with the attributes.

One example for this would be to parse a Ruby date into an Elixir date.